fix any deduping
